adding and editing a document with holes is working
