+++ client/src/utils/markersTools.ts
@@ -1,9 +1,9 @@
 import { Dispatch, SetStateAction } from "react";
 import "../styles/MapComponentsStyles/markers.scss";
-import { createArea, getPolygonCentroid } from "./drawingTools";
 import { Coordinates, Document, fromDocumentTypeToIcon } from "./interfaces";
 import { kirunaCoords } from "./map";
 import { PositionMode } from "./modes";
+import { createArea, getPolygonCentroid } from "./polygonsTools";
 
 export const createMarker = (
   doc: Document,
